Backend auth

+++ backend/api/views/User.py
@@ -20,6 +20,8 @@ from api.models import \
     Profile
 from api.pagination import \
     CustomPagination
+from api.permissions import \
+    IsAdmin
 from api.serializers import \
     UserSerializer
 
@@ -27,11 +29,12 @@ from api.serializers import \
 class UserView(generics.GenericAPIView):
     serializer_class = UserSerializer
     pagination_class = CustomPagination
+    permission_classes = [IsAdmin]
     queryset = User.objects.all().order_by('id')
 
     def get(self, request, *args, **kwargs):
         start_time = time.time()
-        serializer = UserSerializer(self.paginate_queryset(self.queryset), many=True, exclude_fields=['date_joined', 'first_name', 'email', 'last_name', 'is_staff', 'is_active', 'is_superuser', 'last_login', 'groups', 'user_permissions'])
+        serializer = UserSerializer(self.paginate_queryset(self.queryset), many=True, exclude_fields=['date_joined', 'first_name', 'email', 'last_name', 'last_login', 'groups', 'user_permissions'])
         elapsed_time = time.time() - start_time
         print(f"GET USERS: {elapsed_time}")
         return self.get_paginated_response(serializer.data)
@@ -47,6 +50,7 @@ class UserView(generics.GenericAPIView):
 
 class UserDetailView(generics.GenericAPIView):
     serializer_class = UserSerializer
+    permission_classes = [IsAdmin]
     pagination_class = CustomPagination
     queryset = User.objects.all().order_by('id')
 
@@ -61,6 +65,7 @@ class UserDetailView(generics.GenericAPIView):
     def put(self, request, pk, *args, **kwargs):
         try:
             user = User.objects.get(pk=pk)
+            self.check_object_permissions(request, user)
             serializer = self.get_serializer(user, data=request.data)
             if serializer.is_valid():
                 serializer.save()
@@ -73,33 +78,9 @@ class UserDetailView(generics.GenericAPIView):
     def delete(self, request, pk, *args, **kwargs):
         try:
             user = User.objects.get(pk=pk)
+            self.check_object_permissions(request, user)
             user.delete()
             return Response({'message': 'User deleted'}, status=status.HTTP_204_NO_CONTENT)
         except User.DoesNotExist:
             return Response({'error': 'User does not exist'}, status=404)
 
-class GenerateUsers(APIView):
-
-    def get(self, request, *args, **kwargs):
-        n = int(request.GET.get('n'))
-        fake = Faker(['en_AU', 'en_CA', 'en_US', 'es_AR', 'fr_FR', 'es_CO', 'es_ES', 'it_IT', 'de_DE'])
-        fake_en = Faker(['en_AU', 'en_CA', 'en_US'])
-
-        for _ in tqdm(range(n)):
-            fname = fake.first_name()
-            lname = fake.last_name()
-            gender = random.choice(['M', 'F'])
-            bio = ' '.join(fake_en.words(10))
-            age = random.randint(1, 80)
-
-            #create users
-            username = ((fname + lname).lower() + str(random.randint(0,9))).replace(" ", "")
-            while User.objects.filter(username=username):
-                username = ((fname + lname).lower() + str(random.randint(0,9))).replace(" ", "")
-            password = "Password2"
-            user = User.objects.create_user(username=username, password=password, is_active=True)
-
-            #create profile
-            profile = Profile.objects.create(fname=fname, lname=lname, bio=bio, age=age, gender=gender, user=user)
-
-        return Response({"message": f"{n} users created successfully"}, status=status.HTTP_201_CREATED)
